Split AST expression helpers
